-Tweaked SCSP and 68k timings. SCSP should now be correct, 68k still needs a bit of work -Did some minor improvements to the sound mixing code. It's still pretty bad in my opinion -Fixed a bug in Monitor Call Slot/Call Address register that was causing most games using some form of streaming audio(movies, XA audio, etc.) to freeze.
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@342 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/scsp.hh
@@ -57,7 +57,7 @@ private:
   ScspRam *sram;
   SaturnMemory *satmem;
   unsigned long scsptiming1;
-  unsigned long scsptiming2;
+  float scsptiming2;
   FILE *debugfp;
   void Scsp::Convert32to16s(long *, long *, short *, unsigned long);
 public:
